// projects/frontend/src/api/Strategy.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * A strategy actual data
 */
export type Strategy =
  | { "kind": "winner"; "entry": string }
  | {
    "kind": "goals";
    "entry": { country_a_goals: number; country_b_goals: number };
  }
  | { "kind": "cup_winner"; "entry": string }
  | { "kind": "game_to_penalty"; "entry": boolean }
  | { "kind": "first_red_card"; "entry": string }
  | { "kind": "first_yellow_card"; "entry": string }
  | {
    "kind": "penalty_goals";
    "entry": { country_a_goals: number; country_b_goals: number };
  }
  | { "kind": "group_ranking"; "entry": string[] }
  | { "kind": "round_32_qualifiers"; "entry": string[] }
  | { "kind": "round_16_qualifiers"; "entry": string[] }
  | { "kind": "round_8_qualifiers"; "entry": string[] }
  | { "kind": "round_4_qualifiers"; "entry": string[] }
  | { "kind": "third_place_qualifiers"; "entry": string[] }
  | { "kind": "final"; "entry": string[] };